+++ adapters/gnu.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for GNU Prolog 1.4.5 (and later versions)
-%  Last updated on August 23, 2019
+%  Last updated on September 10, 2019
 %
 %  This file is part of Logtalk <https://logtalk.org/>
 %  Copyright 1998-2019 Paulo Moura <pmoura@logtalk.org>
@@ -695,6 +695,22 @@ message_queue_destroy(_) :- fail.
 :- else.
 	'$lgt_user_module_qualification'(Goal, Goal).
 :- endif.
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  auxiliary predicates for compiling modules as objects
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_find_module_predicate'(+atom, -atom, @callable)
+%
+% succeeds when Module:Predicate is visible in module Current
+
+'$lgt_find_visible_module_predicate'(_Current, _Module, _Predicate) :-
+	fail.